feat: Добавлена пагинация

+++ src/main/java/ru/yandex/blog_app/model/dto/Page.java
@@ -21,5 +21,5 @@ public class Page<T> {
 
     private Boolean hasNext;
 
-    private Long lastPage;
+    private Integer lastPage;
 }
